requirements.txt
beautifulsoup4
easyocr
fastapi
Flask
flask_sqlalchemy
google-api-python-client
google-auth-oauthlib
langchain
langchain_core
langchain_together
langchain_community
langchain-mistralai
langchain-ollama
ossapi
pytest-asyncio
pydantic
pymongo
PyPDF2
python-dotenv
redis
Requests
SQLAlchemy
uvicorn
Werkzeug
langchain-openai
langchain-huggingface
pypdf2
faiss-cpu
langdetect
pytest
transformers
datasets
matplotlib
mistralai